requirements.txt
opencv-python
numpy
torch
pillow
onnxruntime